requirements.txt
# ============================================================================
# FILE: requirements.txt
# ============================================================================
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tortoise-orm==0.20.0
asyncpg==0.29.0
python-dotenv==1.0.0
groq==0.11.0
PyPDF2==3.0.1
python-docx==1.1.0
aiofiles==23.2.1
httpx==0.27.0
pandas==2.1.4
openpyxl==3.1.2